move ModuleMetaData to deno_dir.rs
